Tesla alpha 0.6.0

+++ launch_chffrplus.sh
@@ -1,17 +1,28 @@
 #!/usr/bin/bash
 
+export OMP_NUM_THREADS=1
+export MKL_NUM_THREADS=1
+export NUMEXPR_NUM_THREADS=1
+export OPENBLAS_NUM_THREADS=1
+export VECLIB_MAXIMUM_THREADS=1
+
 if [ -z "$PASSIVE" ]; then
   export PASSIVE="1"
 fi
 
+. /data/openpilot/selfdrive/car/tesla/readconfig.sh
+
+
 function launch {
   # apply update
-  if [ "$(git rev-parse HEAD)" != "$(git rev-parse @{u})" ]; then
-     git reset --hard @{u} &&
-     git clean -xdf &&
-     exec "${BASH_SOURCE[0]}"
+  if [ $do_auto_update == "True" ]; then
+    if [ "$(git rev-parse HEAD)" != "$(git rev-parse @{u})" ]; then
+      git reset --hard @{u} &&
+      git clean -xdf &&
+      exec "${BASH_SOURCE[0]}"
+    fi
   fi
-
+  
   # no cpu rationing for now
   echo 0-3 > /dev/cpuset/background/cpus
   echo 0-3 > /dev/cpuset/system-background/cpus
@@ -19,6 +30,8 @@ function launch {
   echo 0-3 > /dev/cpuset/foreground/cpus
   echo 0-3 > /dev/cpuset/android/cpus
 
+  # handle pythonpath
+  ln -s /data/openpilot /data/pythonpath
   export PYTHONPATH="$PWD"
 
   # start manager
